+++ Space Game/Light.h
@@ -8,7 +8,7 @@ public:
 	std::shared_ptr<Entity> parent;
 
 	Light(std::shared_ptr<Entity> parent);
-	bool Update(double deltatime) override;
+	bool Update(float deltatime) override;
 	void Save(std::fstream& f) override { f << (int)nType << " "; };
 	void Load(std::fstream& f) override {};
 };
