+++ Space Game/Orb.h
@@ -7,7 +7,7 @@ class Orb : public Entity
 	float fAge;
 public:
 	Orb(SpaceGame* game, float fX, float fY, float fSpeedX, float fSpeedY);
-	bool Update(double deltatime) override;
+	bool Update(float deltatime) override;
 	bool Collide(Entity* entity) override;
 };
